refactoring film list/grid widgets
