+++ lib/presentation/pages/catalog/pages/top_250.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../widgets/film_grid.dart';
+import '../widgets/film_grid_Impl.dart';
 
 class Top250 extends StatelessWidget {
   const Top250({super.key});
@@ -8,7 +8,7 @@ class Top250 extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      body: const FilmGrid(),
+      body: const FilmGridImpl(),
       backgroundColor: Theme.of(context).colorScheme.onBackground,
     );
   }
