+++ lib/presentation/pages/catalog/pages/top_all.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../widgets/film_list.dart';
+import '../widgets/film_list_Impl.dart';
 
 class TopAll extends StatelessWidget {
   const TopAll({super.key});
@@ -8,7 +8,7 @@ class TopAll extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      body: const FilmList(),
+      body: const FilmListImpl(),
       backgroundColor: Theme.of(context).colorScheme.onBackground,
     );
   }
